# Improves instantiation performance by removing ProxyFunc