… imports/go/example/stars_test.go
… imports/go/gojs.go
… imports/wasi_snapshot_preview1/args_test.go
… imports/wasi_snapshot_preview1/clock_test.go
… imports/wasi_snapshot_preview1/environ_test.go
… imports/wasi_snapshot_preview1/example/README.md
… imports/wasi_snapshot_preview1/fs_test.go
… imports/wasi_snapshot_preview1/poll_test.go
… imports/wasi_snapshot_preview1/wasi.go
… imports/wasi_snapshot_preview1/wrapper.go
… imports/wasi_snapshot_preview1/wrapper_test.go
… internal/gojs/runtime.go
… internal/gojs/spfunc/spfunc.go
… internal/gojs/spfunc/spfunc_test.go
… internal/gojs/syscall.go
… internal/gojs/util.go
… internal/integration_test/vs/bench.go
… internal/wasm/host.go